Voa foguete

+++ src/main/java/br/unitins/tp1/service/Dimensao/DimensaoServiceImpl.java
@@ -5,6 +5,7 @@ import br.unitins.tp1.model.DTO.Televisao.DimensaoResponseDTO;
 import br.unitins.tp1.model.DTO.Televisao.TelevisaoResponseDTO;
 import br.unitins.tp1.model.Televisao.Dimensao;
 import br.unitins.tp1.repository.DimensaoRepository;
+import br.unitins.tp1.validation.ValidationException;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
@@ -19,6 +20,14 @@ public class DimensaoServiceImpl implements DimensaoService{
     @Override
     @Transactional
     public DimensaoResponseDTO create(DimensaoRequestDTO dto) {
+        if (dto == null){
+            throw new ValidationException("DTO", "O dto não pode estar vazio");
+        }
+
+        if (dto.altura() <= 0 || dto.polegada() <= 0 || dto.comprimento() <= 0){
+            throw new ValidationException("DTO", "Escreva apenas valores positivos");
+        }
+
         Dimensao dimensao = new Dimensao();
 
         dimensao.setComprimento(dto.comprimento());
@@ -33,6 +42,18 @@ public class DimensaoServiceImpl implements DimensaoService{
     @Override
     @Transactional
     public void update(long id, DimensaoRequestDTO dto) {
+        if (dto == null){
+            throw new ValidationException("DTO", "O dto não pode estar vazio");
+        }
+
+        if (dto.altura() <= 0 || dto.polegada() <= 0 || dto.comprimento() <= 0){
+            throw new ValidationException("DTO", "Escreva apenas valores positivos");
+        }
+
+        if (dto.altura() == null || dto.polegada() == null || dto.comprimento() == null){
+            throw new ValidationException("DTO", "Não pode conter nenhum valor vazio");
+        }
+
         Dimensao dimensao = dimensaoRepository.findById(id);
 
         dimensao.setComprimento(dto.comprimento());
@@ -43,11 +64,19 @@ public class DimensaoServiceImpl implements DimensaoService{
     @Override
     @Transactional
     public void delete(long id) {
+        if (id <= 0){
+            throw new ValidationException("id", "Escreva apenas valores positivos");
+        }
+
         dimensaoRepository.deleteById(id);
     }
 
     @Override
     public DimensaoResponseDTO findById(long id) {
+        if (id <= 0){
+            throw new ValidationException("id", "Escreva apenas valores positivos");
+        }
+
         return DimensaoResponseDTO.valueOf(dimensaoRepository.findById(id));
     }
 
@@ -58,6 +87,10 @@ public class DimensaoServiceImpl implements DimensaoService{
 
     @Override
     public List<TelevisaoResponseDTO> findTelevisaoByDimensao(long idDimensao) {
+        if (idDimensao <= 0){
+            throw new ValidationException("id", "Escreva apenas valores positivos");
+        }
+
         return dimensaoRepository.findTelevisaoByDimensao(idDimensao).stream().map(TelevisaoResponseDTO::valueOf).toList();
     }
 }
